[ark-tests] feat: add test to ark tests
